+++ src/main/java/com/ayvy/api_java/infrastructure/repository/ClienteRepository.java
@@ -6,13 +6,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
 
-   Optional<Cliente> findByEmail(String email);
+    Optional<Cliente> findByEmail(String email);
 
-   //'Transactional' = se der qualquer erro essa função não é executada
-   // já que é uma operação sensível
-   @Transactional
-   void deleteByEmail(String email);
+    //'Transactional' = se der qualquer erro essa função não é executada
+    // já que é uma operação sensível
+    @Transactional
+    void deleteByEmail(String email);
 
 }
